src/agents/visualizer_agent.py
"""Visualizer Agent - Step 5: Create HTML visualization"""
from agent_framework.azure import AzureAIAgentClient
from models.workflow_schemas import HTMLVisualization


def create_visualizer_agent(client: AzureAIAgentClient):
    """Create the Visualizer Agent"""
    return client.create_agent(
        instructions="""
You are a data visualization specialist for medical decision trees.

**YOUR ROLE:** Convert a decision tree into a diagram JSON format and embed it in HTML.

**INPUT:** Decision tree structure and source_documents from the previous agent

**YOUR TASK:**

1. Transform the decision tree into this JSON format:
   - Create "nodes" array where each decision point and outcome is a node:
     * id: unique number (1, 2, 3...)
     * title: "Rule / Decision Point" for questions, "Assignment" for outcomes
     * content: The question text or risk level assignment
     * type: "code" for outcomes, omit for questions
   
   - Create "connections" array linking nodes:
     * source: id of parent node
     * target: id of child node
     * label: "True" or "False"

2. Generate the complete HTML using this EXACT template:

<!DOCTYPE html>
<html lang="en">
  <head>
    <title>{{IMPAIRMENT_NAME}} Risk Assessment</title>
    <script
      src="https://cdn.scordigital.solutions/components/latest/themes/brms-119d236af7cd.js"
      data-turbo-track="reload"
    ></script>
    <script
      src="https://cdn.scordigital.solutions/components/latest/index.js"
      defer
      data-turbo-track="reload"
    ></script>
    <style>
      body { font-family: Arial, sans-serif; margin: 20px; }
      .sources { background: #f5f5f5; padding: 15px; margin-bottom: 20px; border-radius: 8px; }
      .sources h2 { margin-top: 0; }
      .sources a { display: block; margin: 5px 0; color: #0066cc; }
    </style>
  </head>
  <body>

    <rmv-styles
      asset-path="https://cdn.scordigital.solutions/assets"
    ></rmv-styles>
    <rmv-diagram readonly>
      <rmv-card 
        data-minimap
        variant="overlay" 
        slot="actions"
        hidden
      >
        <rmv-diagram-minimap width="200px" height="150px" style="display: block;"></rmv-diagram-minimap>
      </rmv-card>
      <rmv-row slot="actions" gap="8" grow="false">
        <rmv-card variant="overlay">
          <rmv-column small="0" padding="12">
            <rmv-action-trigger
              text="Sources"
              for="sources-popover"
              variant="tertiary"
            ></rmv-action-trigger>
            <rmv-popover id="sources-popover" hidden position="top-end">
              <rmv-column gap="12" padding="12">
                {{SOURCES_LINKS}}
              </rmv-column>
            </rmv-popover>
          </rmv-column>
          <rmv-row gap="2" padding="4">
            <rmv-button
              variant="tertiary"
              icon="download"
              event-id="toolbar-action"
              value="download"
              style="--button-height: 36px; --button-width: 36px"
            ></rmv-button>
            <rmv-button
              variant="tertiary"
              icon="pip"
              event-id="toolbar-action"
              value="minimap"
              style="--button-height: 36px; --button-width: 36px"
            ></rmv-button>
          </rmv-row>
        </rmv-card>
      </rmv-row>
      <template></template>
    </rmv-diagram>

    <script>
      document.addEventListener("DOMContentLoaded", function () {
        const diagram = document.querySelector("rmv-diagram");
        diagram.data = {{DIAGRAM_JSON}};
      });
    </script>
  </body>
</html>

3. Replace placeholders:
   - {{IMPAIRMENT_NAME}}: the impairment name
   - {{SOURCE_LINKS}}: HTML <rmv-button> tags for each source document with action="<source link"> variant="tertiary" and text="<source title>"
   - {{DIAGRAM_JSON}}: the complete JSON object with nodes and connections

**SOURCE BUTTON TEMPLATE:**
```html
<rmv-button variant="tertiary" text="{{TITLE}}" action="{{URL}}"></rmv-button>
```

Return the complete HTML as a single string in html_content field.
        """,
        name="VisualizerAgent",
        output_schema=HTMLVisualization,
    )
