src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-framework-azure-ai>=1.0.0b251209",
    "agent-framework-core>=1.0.0b251209",
    "agent-framework-devui>=1.0.0b251209",
    "azure-ai-agents>=1.2.0b5",
    "azure-ai-projects>=1.1.0b4",
    "azure-monitor-opentelemetry-exporter>=1.0.0b41",
    "fastapi>=0.124.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
